+++ src/CustomDodgeRoll/CustomDodgeRollItem.cs
@@ -1,10 +1,10 @@
 namespace CwaffingTheGungy;
 
-public interface ICustomDodgeRollItem
+public class CustomDodgeRollItem : PassiveItem
 {
   /// <summary>The CustomDodgeRoll, if any, this item grants while held</summary>
-  public CustomDodgeRoll CustomDodgeRoll();
+  public virtual CustomDodgeRoll CustomDodgeRoll() => null;
 
   /// <summary>The number of extra midair dodge rolls this item grants</summary>
-  public int ExtraMidairDodgeRolls();
+  public virtual int ExtraMidairDodgeRolls() => 0;
 }